moved to pure callback only mode
unittests are deactivated
